requirements.txt
python-dotenv
google-genai
flask
flask_cors
rawpy
imageio